# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Flamoral Dating Platform API
  description: |
    Complete API documentation for the Flamoral Dating Platform.

    This dating platform features:
    - User authentication and profiles
    - Discovery and matching system
    - Real-time messaging
    - Subscription and coins system
    - Boost products
    - Privacy controls
    - Notifications
    - Payment processing
  version: 1.0.0
  contact:
    name: Flamoral Platform Team
    email: platform@flamoral.com
  license:
    name: Proprietary
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://api.flamoral.com
    description: Production API
...
paths:
...
  /api/v1/reports:
    post:
      tags: [Reports]
      summary: Submit report
      operationId: submitReport
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportRequest'
      responses:
        '201':
          description: Report submitted

  # ============== MATCHING SERVICE ==============
...
  /api/v1/notifications:
    get:
      tags: [Notifications]
      summary: Get notifications
      operationId: getNotifications
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Notifications list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedNotifications'
...
  /api/v1/notifications/{id}/read:
    put:
      tags: [Notifications]
      summary: Mark notification as read
      operationId: markNotificationRead
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Marked as read
  /api/v1/notifications/read-all:
    put:
      tags: [Notifications]
      summary: Mark all notifications as read
      operationId: markAllNotificationsRead
      security:
        - bearerAuth: []
      responses:
        '200':
          description: All marked as read
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    ReportRequest:
      type: object
      required: [userId, categoryId]
      properties:
        userId:
          type: string
        categoryId:
          type: string
        description:
          type: string
...
    Notification:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
        title:
          type: string
        body:
          type: string
        read:
          type: boolean
        createdAt:
          type: string
          format: date-time
    PaginatedNotifications:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Notification'
        pagination:
          $ref: '#/components/schemas/Pagination'
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer